Fix : 전체적인 블로그 로직(좋아요, 서비스 필드 등) 수정 및 응답 메시지 추가

+++ src/main/java/com/myalley/blogReview/service/BlogImageService.java
@@ -5,14 +5,13 @@ import com.myalley.blogReview.domain.BlogReview;
 import com.myalley.blogReview.repository.BlogImageRepository;
 import com.myalley.exception.BlogReviewExceptionType;
 import com.myalley.exception.CustomException;
+import com.myalley.member.domain.Member;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.util.CollectionUtils;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
-import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
 
 @Service
@@ -22,20 +21,17 @@ public class BlogImageService {
     private final S3Service s3Service;
 
     public void addBlogImageList(List<MultipartFile> images, BlogReview blogReview) throws IOException {
-        //HashMap<String,String> map = s3Service.uploadBlogImages(images);
         if (!CollectionUtils.isEmpty(images)) {
             for (MultipartFile imageFile : images) {
                 String[] imageInformation = s3Service.uploadBlogImage(imageFile);
                 addBlogImage(imageInformation[0],imageInformation[1],blogReview);
-                //imageInformationMaps.put(imageInformation[0], imageInformation[1]);
             }
         }
-        //map.forEach((fileName,S3url)->{
-       //     addBlogImage(fileName,S3url,blogReview);
-        //});
     }
 
-    public void createNewBlogImage(BlogReview blogReview, MultipartFile image) throws IOException {
+    public void createNewBlogImage(BlogReview blogReview, Member member, MultipartFile image) throws IOException {
+        if(blogReview.getMember().getMemberId()!= member.getMemberId())
+            throw new CustomException(BlogReviewExceptionType.IMAGE_FORBIDDEN);
         String[] information = s3Service.uploadBlogImage(image);
         addBlogImage(information[0],information[1],blogReview);
     }
@@ -49,7 +45,9 @@ public class BlogImageService {
         blogImageRepository.save(newImage);
     }
 
-    public void removeBlogImage(BlogReview blogReview,Long imageId){
+    public void removeBlogImage(BlogReview blogReview, Member member, Long imageId){
+        if(blogReview.getMember().getMemberId()!= member.getMemberId())
+            throw new CustomException(BlogReviewExceptionType.IMAGE_FORBIDDEN);
         BlogImage foundImage = retrieveBlogImage(blogReview,imageId);
         s3Service.deleteBlogImage(foundImage.getFileName());
         blogImageRepository.delete(foundImage);
@@ -58,14 +56,10 @@ public class BlogImageService {
     public void removeBlogAllImages(BlogReview blogReview) {
         List<BlogImage> blogImageList = blogImageRepository.findAllByBlog(blogReview);
         if (!CollectionUtils.isEmpty(blogImageList)) {
-            //List<String> fileNames = new ArrayList<>();
             for (BlogImage blogImage : blogImageList) {
                 blogImageRepository.delete(blogImage);
                 s3Service.deleteBlogImage(blogImage.getFileName());
-                //fileNames.add(blogImage.getFileName());
             }
-            //if(!CollectionUtils.isEmpty(fileNames))
-            //    s3Service.deleteBlogAllImages(fileNames);
         }
     }
 
@@ -75,15 +69,4 @@ public class BlogImageService {
         });
         return image;
     }
-
-    /*
-    public List<BlogImage> retrieveBlogImages(Long blogId){
-        //blogReviewService
-        List<BlogImage> images = blogImageRepository.findAllByBlog(blogId);
-        if(images.isEmpty())
-            throw new CustomException(BlogReviewExceptionType.IMAGE_NOT_FOUND);
-        return images;
-    }
-
- */
 }
